+++ storage/connect/jdbconn.cpp
@@ -1215,7 +1215,7 @@ void JDBConn::SetColumnValue(int rank, PSZ name, PVAL val)
 	if (rank == 0)
 		if (!name || (jn = env->NewStringUTF(name)) == nullptr) {
 			sprintf(g->Message, "Fail to allocate jstring %s", SVP(name));
-			throw TYPE_AM_JDBC;
+			throw (int)TYPE_AM_JDBC;
 		}	// endif name
 
 	// Returns 666 is case of error
@@ -1223,7 +1223,7 @@ void JDBConn::SetColumnValue(int rank, PSZ name, PVAL val)
 
 	if (Check((ctyp == 666) ? -1 : 1)) {
 		sprintf(g->Message, "Getting ctyp: %s", Msg);
-		throw TYPE_AM_JDBC;
+		throw (int)TYPE_AM_JDBC;
 	} // endif Check
 
 	if (val->GetNullable())
@@ -1334,7 +1334,7 @@ void JDBConn::SetColumnValue(int rank, PSZ name, PVAL val)
 			env->DeleteLocalRef(jn);
 
 		sprintf(g->Message, "SetColumnValue: %s rank=%d ctyp=%d", Msg, rank, (int)ctyp);
-		throw TYPE_AM_JDBC;
+		throw (int)TYPE_AM_JDBC;
 	} // endif Check
 
 	if (rank == 0)
